move function that gets primary key names to column util class
